+++ packages/cli-plugin-package/src/index.ts
@@ -177,15 +177,6 @@ export class PackagePlugin extends BasePlugin {
         this.options.sourceDir
       );
     }
-    // 分析midway version
-
-    const faasModulePath = findNpmModule(cwd, '@midwayjs/faas');
-    if (faasModulePath) {
-      const pkgJson = JSON.parse(
-        readFileSync(join(faasModulePath, 'package.json')).toString()
-      );
-      this.midwayVersion = pkgJson.version[0];
-    }
 
     // 分析目录结构
     const locator = new Locator(this.servicePath);
@@ -250,6 +241,17 @@ export class PackagePlugin extends BasePlugin {
     } else {
       this.core.cli.log(' - Find node_modules and skip...');
     }
+
+    // 分析midway version
+    const cwd = this.getCwd();
+    const faasModulePath = findNpmModule(cwd, '@midwayjs/faas');
+    if (faasModulePath) {
+      const pkgJson = JSON.parse(
+        readFileSync(join(faasModulePath, 'package.json')).toString()
+      );
+      this.midwayVersion = pkgJson.version[0];
+    }
+    this.core.debug('midwayVersion', this.midwayVersion);
   }
 
   async copyFile() {
@@ -548,16 +550,17 @@ export class PackagePlugin extends BasePlugin {
     }
     // midway 2版本的装饰器分析由框架提供了
     if (this.midwayVersion === '2') {
-      process.chdir(this.midwayBuildPath);
-      const httpFuncSpec = await analysisDecorator(
-        join(this.midwayBuildPath, 'dist')
-      );
-      process.chdir(this.core.cwd);
       if (!this.core.service.functions) {
         this.core.service.functions = {};
       }
-      this.core.debug('httpFuncSpec', httpFuncSpec);
-      Object.assign(this.core.service.functions, httpFuncSpec);
+      process.chdir(this.midwayBuildPath);
+      const funcSpec = await analysisDecorator(
+        join(this.midwayBuildPath, 'dist'),
+        this.core.service.functions
+      );
+      process.chdir(this.core.cwd);
+      this.core.service.functions = funcSpec;
+      this.core.debug('funcSpec', funcSpec);
     }
   }
 
@@ -895,7 +898,7 @@ export class PackagePlugin extends BasePlugin {
         console.error(
           `Cannot use the same prefix '${currentPath}' for aggregation deployment`
         );
-        process.exit();
+        process.exit(1);
       }
       allAggregationPaths.push(currentPath);
       this.core.service.functions[aggregationFuncName]._handlers = handlers;