+++ packages/cli-plugin-faas/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.2.59](https://github.com/midwayjs/cli/compare/v1.2.58...v1.2.59) (2021-04-07)
+
+**Note:** Version bump only for package @midwayjs/cli-plugin-faas
+
+
+
+
+
+## [1.2.58](https://github.com/midwayjs/cli/compare/v1.2.57...v1.2.58) (2021-04-07)
+
+**Note:** Version bump only for package @midwayjs/cli-plugin-faas
+
+
+
+
+
+## [1.2.57](https://github.com/midwayjs/cli/compare/v1.2.56...v1.2.57) (2021-04-02)
+
+**Note:** Version bump only for package @midwayjs/cli-plugin-faas
+
+
+
+
+
 ## [1.2.56](https://github.com/midwayjs/cli/compare/v1.2.55...v1.2.56) (2021-03-26)
 
 **Note:** Version bump only for package @midwayjs/cli-plugin-faas